--- a/src/doxygen.cpp
+++ b/src/doxygen.cpp
@@ -11922,12 +11922,9 @@ void generateOutput()
     g_s.end();
   }
 
-  if (Config_getBool(PLANTUML_RUN_FAST))
-  {
-    g_s.begin("Running plantuml with JAVA...\n");
-    PlantumlManager::instance()->run();
-    g_s.end();
-  }
+  g_s.begin("Running plantuml with JAVA...\n");
+  PlantumlManager::instance()->run();
+  g_s.end();
 
   if (Config_getBool(HAVE_DOT))
   {