@mosra 2019-01-03
A friendly suggestion :wink: : you don't have to close pull requests and open new ones with updates, it's enough to just `git push` (or `git push -f`) to the original one. That will keep all the comment history and context for easier reviewing and you won't need to repeat everything every time.

Besides that (excuse me if I missed something), but why have an option for `PLANT_RUN_FAST`, why not make that the default? I don't see an use case where the user would want to explicitly run it slow. Having two code paths also means there is more code to test and more code that can have bugs. If I remember correctly, this optimization is done for LaTeX formulas as well, they're all being processed at once at the end, and there's no option for `LATEX_RUN_FAST` either.