@albert-github 2019-01-10
@cheoljoo  When I don't have any plantuml diagrams in my code I still get (in this case) empty `inline_umlgraph_cache_all.pu`. I don't think this file should be created in this case or at least removed at the end of the run.