@mosra 2019-01-10
Yeah, I'm getting that too, was about to open an issue :) Thanks for pointing that out!