@albert-github 2019-01-10
Code of #6762 has been integrated in master on github.